+++ include/ignition/gazebo/components/JointVelocityReset.hh
@@ -14,13 +14,13 @@
  * limitations under the License.
  *
  */
-#ifndef IGNITION_GAZEBO_COMPONENTS_JOINTPOSITION_HH_
-#define IGNITION_GAZEBO_COMPONENTS_JOINTPOSITION_HH_
+#ifndef IGNITION_GAZEBO_COMPONENTS_JOINTVELOCITYRESET_HH_
+#define IGNITION_GAZEBO_COMPONENTS_JOINTVELOCITYRESET_HH_
 
 #include <vector>
 
-#include <ignition/gazebo/components/Factory.hh>
 #include <ignition/gazebo/components/Component.hh>
+#include <ignition/gazebo/components/Factory.hh>
 #include <ignition/gazebo/components/Serialization.hh>
 #include <ignition/gazebo/config.hh>
 
@@ -32,13 +32,16 @@ namespace gazebo
 inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
 namespace components
 {
-  /// \brief Joint positions in SI units (rad for revolute, m for prismatic).
+  /// \brief Joint velocities in SI units
+  ///        (rad/s for revolute, m/s for prismatic).
+  ///
   /// The component wraps a std::vector of size equal to the degrees of freedom
   /// of the joint.
-  using JointPosition = Component<std::vector<double>, class JointPositionTag,
-                                  serializers::VectorDoubleSerializer>;
+  using JointVelocityReset = Component<std::vector<double>,
+                                       class JointVelocityResetTag,
+                                       serializers::VectorDoubleSerializer>;
   IGN_GAZEBO_REGISTER_COMPONENT(
-      "ign_gazebo_components.JointPosition", JointPosition)
+      "ign_gazebo_components.JointVelocityReset", JointVelocityReset)
 }
 }
 }