+++ src/gui/plugins/grid_config/GridConfig.cc
@@ -87,7 +87,7 @@ void GridConfig::LoadConfig(const tinyxml2::XMLElement *)
 /////////////////////////////////////////////////
 bool GridConfig::eventFilter(QObject *_obj, QEvent *_event)
 {
-  if (_event->type() == ignition::gazebo::gui::events::Render::Type)
+  if (_event->type() == ignition::gazebo::gui::events::Render::kType)
   {
     // This event is called in Scene3d's RenderThread, so it's safe to make
     // rendering calls here
@@ -101,15 +101,13 @@ bool GridConfig::eventFilter(QObject *_obj, QEvent *_event)
 /////////////////////////////////////////////////
 void GridConfig::UpdateGrid()
 {
-  if (nullptr == this->dataPtr->grid)
-  {
+  // Load grid if it doesn't already exist
+  if (!this->dataPtr->grid)
     this->LoadGrid();
-  }
 
-  if (nullptr == this->dataPtr->grid)
-  {
+  // If grid was not loaded successfully, don't update
+  if (!this->dataPtr->grid)
     return;
-  }
 
   if (!this->dataPtr->dirty)
     return;
@@ -134,7 +132,7 @@ void GridConfig::UpdateGrid()
     visual->SetVisible(this->dataPtr->gridParam.visible);
   }
 
-  this->dataPtr->dirty = true;
+  this->dataPtr->dirty = false;
 }
 
 /////////////////////////////////////////////////